feat(coach): migration to typescript and fix review-request to review
