--- lib/pages/coach/coach-review-index/components/index.jsx
@@ -1,4 +1,0 @@
-import Page from './Page'
-import CoachReviewCard from './CoachReviewCard'
-
-export { Page, CoachReviewCard }
