+++ lib/pages/coach/coach-review-index/components/index.tsx
@@ -1,0 +1,4 @@
+import Page from "./Page"
+import CoachReviewCard from "./CoachReviewCard"
+
+export { Page, CoachReviewCard }
