--- lib/pages/coach/coach-review/components/index.jsx
@@ -1,7 +1,0 @@
-import Page from './Page'
-import CommentBar from './CommentBar'
-import CommentForm from './CommentForm'
-
-const commentDuration = 10
-
-export { Page, CommentBar, commentDuration, CommentForm }
