--- pages/coach-reviews/[id].js
@@ -1,3 +1,0 @@
-import { Page } from '~/pages/coach/coach-review/components'
-
-export default Page
