+++ pages/coach-reviews/[id].ts
@@ -1,0 +1,3 @@
+import { Page } from "gamejitsu/pages/coach/coach-review/components"
+
+export default Page
